refactor : 기존 running/battle 한 곳에 몰았던 파일들을 디렉토리별 목적에 맞게 분리
